+++ docs/md/releasenotes.md
@@ -9,6 +9,7 @@
 ### Bug fixes
 
 - Fix handling of category 2 values in `JvmValueTransferRelation` to work correctly with `ExecutingInvocationUnit`.
+- Fix concurrency problems in CallGraph and ValueFactory ids
 
 ## Version 9.1.1